capture/requirements.txt
opencv-python-headless==4.9.0.80
numpy==1.26.4
pyzmq==25.1.2
